Remove temporary issue-gathering button EN, NL (#12)
This change removes the temporary button for gathering work issues from
the homepage in both the EN and NL versions. This button had been added
for the MozFest event in early June. As this event has now passed, we
can remove this button again -- as discussed in this [Slack
conversation](https://techworkersco.slack.com/archives/C06NYM3271D/p1719557456287469?thread_ts=1719305663.184519&cid=C06NYM3271D).

Btw it was also pointed out that the blurb generally could be better,
which is totally valid! This is just to make this specific, limited fix.