> throw new Error('Specified "i18n" cannot be used with "output: export". See more info here: https://nextjs.org/docs/messages/export-no-i18n');
Whelp, this is the reason we can't use NextJS I think. We want SSG, and
we "need" it on netlify (forms integration which does slack sign up and
email subscription things). NextJS would require us to either pay for
vercel (lose netlify things), or run our own server (lose netlify + have
to deal with maintainence).

With that in mind, we go Gatsby!