Make sure all links contain a locale
This is opinionated. It's mainly to make things consistent so it's
harder to miss things when developing. e.g. developers often don't test
sites in multi languages. It is also harder to cater for the site url
diff when dealing with multiple languages (e.g. js vs go vs. hugo in go)