Implement Decap CMS for managing resources/events (#62)
The diff looks scarier than it should, explanation:

- content/events - have adjusted all them to be `index.<locale>.md`
which makes them show up pretty too.
- Some changes look like delete/add, fear not, the diff was just enough
that git thought it wasn't a change
- static/admin - decap itself and configurations
- layouts/* - Using the default `.Date` field everywhere instead
- events_list change is mostly whitespace


A note on decap: The folders that it creates for the events are the date
of the event creation. It's just weird and unavoidable decap things. I
don't think this will matter in the real world, but if it does, then
that's slightly painful. Let's see what it does, we might need some sort
of github bot to adjust and throw a commit onto the branch or something
(I'm happy to play around with this if this is the case).


Other notes:
- event.description was only filled in one place, and we don't display
it anywhere
- event.event_date superseeded by date
- event.event_time superseeded by date
- removed a bunch of checks for if date was nil or not. IMO it will
always be provided on events
- the article pagination is back to what it was in congo I think
- a smol fix in the events_list where it wasn't generating a closing div
inside the range

Note: By default, Hugo does not build any content where the `date` field
is in the future. I guess the idea is to not publish things that are not
yet ready for the real world. I changed this behaviour by putting
setting the `date` field on events to be the event date, rather than the
"publishing date" (which i guess is the intended use case). We can
always change this back in the future, but for now it's one less field.

## Tests

`hugo server` in one terminal
`npx decap-server` in another

And then browse to http://localhost:1313/admin

Note: It also works on the deployment preview site:
https://deploy-preview-62--twc-site-nl.netlify.app/admin <-- Beware, it
is using the content in `main`, not what is on the branch