Add events april 2025 (#78)
This change adds a couple of new events to the events list:
* Organizing meetups on 7 April, 21 April, and 19 May
* Tech workers conference Berlin external event on 11 April
* TWC 101 onboarding call on 15 April

It also shortens the titles for the Organizing for Power past events,
purely for aesthetic reasons, and simplifies the blurb on the events
landing page.

Gonna merge this without review as it's mainly adding a couple of
recurring events, and events of which the content has already been
determined elsewhere.

---------

Co-authored-by: Kit <204160306+kitonrust@users.noreply.github.com>