Add flyer to booking event (#82)
This change adds a flyer to the Booking info sessions. In addition:
* It adds a description to those events. `description` is currently
missing from our events template, but it's needed for meta info in the
page `head`, to help with SEO 🙃 Just added it for these two events for
now, but let's add this an all pages going forward.
* Also shrinks the file size of another events image, just to help keep
pages load fast.

---------

Co-authored-by: Kit <204160306+kitonrust@users.noreply.github.com>