Add tech workers nl 4 palestine section (#76)
This change:

## Add campaigns section

- Adds a section **Campaigns**, which can list special focus topics
around which we might want to create multiple resources. Currently (with
this PR) it'll include a single campaign, but I can image we might also
have one around AI or workplace rights or smth? Next steps flexible 🙂
This includes:
- Adds a `campaigns-list` layout, which uses `campaign-link` instead of
`article-link`. The main (only) difference with this layout is that it
won't list the date an article was published, which seemed to me a
distraction.
- The campaigns list also shows the campaign summary, as long as
`showSummary` is set to `true` in the campaign's frontmatter.
- The choice to calls this **Campaigns**, rather e.g. Topics or anything
else, was [copied from UTAW](https://utaw.tech/campaigns), who also use
this terminology.

## Add campaign: Tech workers for Palestine
- Adds a campaign 'Tech workers for Palestine' (EN/NL) to the campaigns
section, including feature image.

## Add campaign single page

- Adds a layout `campaign` for the landing page of an individual
campaign. The main difference with a default single article layout is
that this loads a bunch of resources related to the campaign at the
bottom. Which resources get listed is determined based on the following:
- To the resources that you want to list, you manually add a tag that
identifies the relevant campaign. In this case: `palestine`.
- Then in the frontmatter of the landing page of the campaign itself,
you specify `campaign` with the tag that identifies the campaign. In
this case: `campaign: "palestine"`.

## Other related changes

- Hide the tag list on the resources link. Motivation: The styling
looked (imho) not great, and we're currently not really using or showing
tags systematically. So I'd say until we've decided how we'd like to use
and show tags, let's hide it for now.
- Including this change in the current PR, as otherwise the tag was also
showing up when listing resources at the bottom of the new campaign
page.

![Screenshot 2025-03-23 at 15 12
50](https://github.com/user-attachments/assets/7cc14aad-0aa7-4372-99ce-d4b6985e2e3b)
*Tag on resource. At this point it doesn't look great imo, so propose to
hide it for now*

- Update the cover image on the *Organize your workplace* resource. The
old one was just too cringe 😅 The stock photo images obvs still aren't
ideal, so as soon as anyone steps in with something better let's go for
it 😁
- Including this change in the current PR, because I'd like to list the
'Organize your workplace' as a resource at the bottom of the 'Tech
workers for Palestine' landing page, and otherwise it still had the
older image

![Screenshot 2025-03-23 at 15 17
12](https://github.com/user-attachments/assets/40286ebe-e1da-4e84-84ad-91d0af663d46)
*Swapped feature image on Organize your workplace resourece*

---------

Co-authored-by: Kit <204160306+kitonrust@users.noreply.github.com>
Co-authored-by: Ben Naylor <nayben@gmail.com>