Use different backend for CMS (#88)
git-gateway looks deprecated by netlify:
https://docs.netlify.com/security/secure-access-to-sites/git-gateway/

The way it worked: We stored a PAT token in netlify. When people logged
into the CMS, they were logging into netlify. Netlify would then use
that PAT token to commit the changes. This meant we could give people
access to netlify, instead of the git repo. It also meant people didn't
have to have a git account. Reality is that the editing group likely
isn't large, and we can get those people to create github accounts.

Also: Enable open authoring. Forks and commits PRs for anyone:
https://decapcms.org/docs/open-authoring/

:point_up: This will mean anyone can go to https://techwerkers.nl/admin,
and they will be able to create new records. It will fork the repo into
their github account, make the changes, and PR in the background - very
cool. I doubt we'll "need" that so much, but it's a "nice to have".