Adjust hugo layouts folders (#91)
I discovered this recently in another PR:
https://gohugo.io/templates/lookup-order/#single-templates

In short:

- Content is stored `content/<type>/<slug>/index.<locale>.md`
- Layouts are looked up in `layouts/<type>/single.html` (or `list.html`
for `_index.<locale>.md`)

If a relevant `single.html` (or `list.html`) file isn't found, the it
goes to `layouts/_defaults/single.html` (or a name based on the `layout
attribute in the frontmatter`.

I realised we were adding `layout` in lots of places to do the search in
the _default folder, however we can just use/abuse the paths to get our
files instead :smile:

Notes:
- `events`, `campaigns`, `resources` all follow this `list` + `single`
pattern. I just moved their relevant layouts out etc.
- `join` as a little different, but the adjustments here all work. both
`/join` and `/join/success` are single pages, but I thought it looks
easier to understand that `success` still has a layout of it's own.

And one sadish note - without having a `layout` field on each resource,
there is no way to filter out the main list page for each type in the
CMS. It's both sad, but maybe also kinda good. It means it is possible
to edit the list page in the CMS now too (maybe handy?).

## Notes for review

- All changes in `content/<type>` are all the same. Removing the layout
field. I did this with a regex, and am pretty sure it caught everything
:crossed_fingers:.
- Content was moved from layouts/_default -> layouts/<type> so it should
be exactly the same :crossed_fingers:

This is how the layouts directory structure now looks:


![Screenshot_20250501_093451](https://github.com/user-attachments/assets/4f0f87e5-73ac-4c7f-bdc8-5580958fa4c9)

## Notes on CMS

I tried to make the index pages in the CMS a little more visible:

| Events | Resources |
| - | - |
|
![Screenshot_20250501_095618](https://github.com/user-attachments/assets/b8225a84-ffd1-4fdd-8fb7-a29cd060ad6e)
|
![Screenshot_20250501_095611](https://github.com/user-attachments/assets/a3cb3b17-6157-4b5d-8cc1-cfc455ed46e6)
|

That was done in this commit:
https://github.com/techworkersco/twc-site-nl/pull/91/commits/ae7bdedd8be3b79f80b2faea2964a00599c8ec19
(see commit info for more info, TL;DR: these are limited go templates
sadly)