Add events.yaml for global site (#90)
Lots of "TIL" for me here.

Hugo wants to have output formats "per content". e.g. we have a main
_index collection page for events, and current we render this in HTML
(and also JSON, and RSS xml aswell!)

So what we "needed" here is a new output format for yaml - and then
telling the events page we also want to have a YAML output. We could
define yaml output for the entire site, however hugo starts to complain
since we don't have a global yaml. Fun facts:
https://techwerkers.nl/en/index.xml <-- we have a RSS feed for our
homepage at present :sweat_smile:

So this PR:
- [x] adds a new yaml output format
- [x] tells the event page to output a yaml page as well as html and rss
(which we already have)
- [x] includes a specific yaml layout for the events page (which makes
sense because it's different frontmatter etc)

And bonuses:
- [x] fixes some event timezones, they should have been +02:00
- [x] adds CMS feature image support :rocket: 

:brain: At some point in the future we could enable an `ics` output
format. Then we could actually render an `ics` (calendar) file for each
event, and put a link on the page for people to add it to their
calendars :sweat_smile: (The hard part is we only want `ics` for each
event, and that means adding an output array for each event, maybe some
hacky ways to do this, who knows).

I also just stumbled across this:
https://discourse.gohugo.io/t/custom-layout-for-certain-content-types/3193
which might be an easier way to do layouts per content type :shrug:.