Ignore content for prettier, decap and prettier don't play togethermn (#97)
There isn't any way to control the output of the content from decap, and
it all "Just works" normally. And it looks like there isn't a way to
configure decap output frontmatter to be prettier compliant, so :shrug:.

Alternative is to get prettier to just run from a github action and
commit itself, but meh? It's a bit of messing around for not much
benefit, so if someone else is keen to make that work, all good.