Merge congo into our repo, no more dependency (#107)
Fixes: #106 

This PR merges all of the upstream jpanther/congo template into this
repo. We make a number of small changes which aren't always compatible
with the upstream (specifically for our style/content design/layouts).
We may want to vary even more in the future, so this is mainlining the
repo, and letting us deviate into our own special template etc.

The thought is that we won't need many changes from the mainline congo
template anymore, and that we're happy with merging some dependabots
every now and then (I'll set up npm dependabot in another PR). We can
always carve our changes out again if we want to extend the main congo
template (although that'll be hard no doubt).