(Fix) Clarify no bosses or managers on join page (#100)
## Update (2025-05-15):

Made some changes, might be overkill 😅 Changes:

- Added some text to clarify who can & can't join, inspired by:
  - https://techworkerscoalition.org/subscribe/
  - https://techworkerscoalition.org/community-guide/
- However, instead of adding it either at the top or bottom of the Join
page (where I tried it initially), added it on a separate `principles`
(short for 'joining principles' or whatever) page, so that it doesn't
create a negative mood on the Join page.
- Adapted the join layout so that the feature image gets loaded
automatically, instead of having to be added manually.

Known issue:
- [ ] _Something_ is overruling the styling on this page, but I can't
for the life of me figure out what's doing it. This makes it the case
that the link to the eligibility principles page doesn't show up as a
link, and that the submit button lacks the styling of the regular submit
button. Would be great if this could be fixed somehow :)

## Earlier (2025-05-13):

This change adds a note on the Join page to clarify that you can't join
if you've got hiring/firing/promotion/demotion power over other workers
--- i.e. no bosses or people managers.

Previously, we had a similar note on this page, but it perished in the
Grand Join Page Cleanup a couple of months ago. This change adds such a
note back in.

In addition:
- It removes 'we' from the join description, and simplifies this
description a bit.
- On the Dutch page, it changes the heading with a CTA to join
'Techwerkers', rather than 'Techwerkerscoalitie', which is just a bit of
a mouthful 😄

---------

Co-authored-by: Kit <204160306+kitonrust@users.noreply.github.com>
Co-authored-by: Ben Naylor <nayben@gmail.com>