Isolate twc-global specific layouts and content (#112)
Use the content/layouts directory for things that twc-global
specifically needs. At present it's just the events page. This way
customizations can be put there and we don't need to think too much
about the "why" etc.

Why this PR:

- #90 added an events yaml for the global site consumption
- #108 #111 adds a prefix to the event title <-- specifically all the
events on the twc-site-nl are for Nederland, so having this prefix in
this folder is a twc-global concern etc.

This is probably over engineering things a little :sweat_smile:. One
thing we could eventually do is to include other sections into the
twc-global folder - specifically for consuming posts if that wanted to
be a thing too etc.