Add news item support for railway workers (#117)
This change adds a news post that expresses TWC NL's support for railway
workers' workplace action to get better wages & working conditions.

Support discussed on Slack. 

The post image and content are identical to what's already posted on our
socials, and thought it might be nice to also have this on our own site
as well.

Co-authored-by: Kit <204160306+kitonrust@users.noreply.github.com>