Layout and config tweaks (#135)
To be checked/merged after #132 

Changes in this PR:
- Adding about page and github link in the global footer
- Removing sharing links from the about page - there was a lot of icons,
not sure if it's entirely necessary etc.
- Include some root domain routes for about (e.g. techwerkers.nl/about)
- Enable key value pair i18n. This is handy for the underlying congo
template for showing dates/times and when things were last updated - I
broke this when merging the template in apparently :sweat_smile:
- Make the resources a sort of "type" which is handy for configuring
some of the visibility of things. (this is the params.toml changes)
- Center all the pages
- Put the join page content into the content page itself. This makes it
easier to track the text changes over time

Future:
- I think it's possible to dedupe some of the format/layout on some of
our custom pages. Might make things a little easier to manage there (not
that we ever need to, but you know what I mean)
- key value pair for the page content - e.g. "join" fields etc could be
in a `join.title` and `join.form.name` and `join.form.email` etc.

---------

Co-authored-by: Kit <204160306+kitonrust@users.noreply.github.com>