Adding some animated TWC logos (#136)
Why?

Answer: Yes

I'm not sure if we "need" these, as much as it's fun to have them.

Process:

1. svgomg.net on the original svg rendered assets to reduce the
precision to 2dp, and to remove some of the superfluous inkscape stuff.
2. Pull apart all the SVG elements, group them, give them sensible Ids,
manually round a bunch of the elements
3. Rewrite a whole bunch of the SVG paths by hand by doing some math.
FYI, the lines in the bottom have a 60% angle, e.g. for every 3x,
increment 5y etc.
4. Write a bunch of animations from scratch too, and mess with their
timings to there's some call and response going on.
5. iterate, iterate, iterate, iterate - hand editing SVGs is pretty fun
though!