Add all the infra things that were needed for docmost (#144)
This PR includes all the setup that was needed for docmost.

Added a small amount of docs about how to execute this stuff.

Note: The secrets are fine to be committed because they are encrypted
(as in proper encryption).

The additions here require access to the keepass in order be able to do
a lot of the things here. In an ideal world we'd put a lot of this stuff
in a separate repo. But for now this works :sweat_smile: