Snapshot backups (#160)
The previous style of backups were pretty naive without any cleanup.
e.g. it chews through disk space. Not a worry at present, but could be
an issue later.

I was looking for a solution for my selfhosted immich, and came across
rsnapshot which is pretty nice. It does a lot of deduplicating, and
keeps track of changes over time. I had a look for other kubernetes
native things, however they're all quite heavy and/or would result in
more spending etc.

The main changes are:
- Use `rsnapshot` for doing the snapshot of data directories
- Use SSHFS to mount the existing backups so rsnapshot can send them
places. This is outside the normal design of rsnapshot from what I was
reading, but it also "works".
- Change a bunch of the logging to make it easier to debug. There was
alot of debug
- A snapshot weekly with 4 weeks retention, snapshot daily for the last
week.

Some other notes:
- Since the psql is just a blob of SQL, there is no snapshot/incremental
diff here. There was some cool new things in postgres 17, but I didn't
want to experiment with them yet. Also, the DB size is 1.5Mb here, where
as the data volume for docmost is 70MB, so that was the more pressing
thing.

TODO (will be done in other PRs/later)
- [ ] Migrate nocodb to also use the same snapshotting
- [ ] Clean up the old backups